A dual server/client example using Lattice
**Overview**

The `lattice-markdown-server-client` boilerplate project allows you
an easy starting place with which to work with an app that is both
react on the client and node/graphql/graphql-lattice on the server.

A *very* simplistic approach to reading markdown content off the disk on
the server and serving it via graphql is setup on the server. The server
endpoint is automagically translated via the `"proxy"` field in the
package.json file and understood by the create-react-app client.

See (this
site)[https://daveceddia.com/create-react-app-express-backend/] for more
information on how the proxy situation works.

**Moving Parts**

The boilerplate is a combination of two other types plus a very minimal
amount of code. It uses `lattice-quickstart` for the server piece and
`create-react-app` generated code for the client. The former was cloned
into a subdirectory called `server` and then its dependencies were
merged with those of the react app.

ApolloClient is used on the react side of things and this starting point
is left as an example of how to do the same sort of work on your own or
as a base from which to build upon.